+++ calcipy/doit_tasks/__init__.py
@@ -29,6 +29,7 @@ __all__ = [  # noqa: F405
     'task_pre_commit_hooks',
     'task_radon_lint',
     'task_security_checks',
+    'task_static_checks',
     # from .packaging
     'task_check_for_stale_packages',
     'task_check_license',
@@ -77,6 +78,7 @@ TASKS_LOCAL = [
     'check_for_stale_packages',
     'pre_commit_hooks',
     'lint_project',
+    'static_checks',
     'security_checks',
     'check_types',
 ]