+++ calcipy/file_helpers.py
@@ -57,9 +57,9 @@ def _read_yaml_file(path_yaml: Path) -> Any:
     # TODO: modify so that mkdocs.yml can be read, but Python won't be executed...
 
     # Based on: https://github.com/yaml/pyyaml/issues/86#issuecomment-380252434
-    yaml.add_multi_constructor('', lambda loader, suffix, node: None)
-    yaml.add_multi_constructor('!', lambda loader, suffix, node: None)
-    yaml.add_multi_constructor('!!', lambda loader, suffix, node: None)
+    yaml.add_multi_constructor('', lambda _loader, _suffix, _node: None)
+    yaml.add_multi_constructor('!', lambda _loader, _suffix, _node: None)
+    yaml.add_multi_constructor('!!', lambda _loader, _suffix, _node: None)
     try:
         return yaml.unsafe_load(path_yaml.read_text())
     except (FileNotFoundError, KeyError) as err:  # pragma: no cover