+++ calcipy/log_helpers.py
@@ -1,5 +1,14 @@
 """Loguru Helpers."""
 
+# PLANNED: consider a STDOUT format like https://pypi.org/project/readable-log-formatter
+#   COlorful Debug Level / Parent/FileName / Line NUmber
+#       Indented and wrapped summary string
+#       Indented, optional variables (would otherwise not be shown) (i.e. **{name}**: {value})
+#   Note: need to ensure that exception tracebacks are properly handled with a custom format
+#       See: https://github.com/Delgan/loguru/issues/156
+#           And: https://loguru.readthedocs.io/en/stable/api/logger.html#message
+#       Or: https://github.com/Delgan/loguru/issues/133
+
 from __future__ import annotations
 
 import logging
@@ -163,7 +172,7 @@ def _format_jsonl_handler(log_dir: Path, production: bool = True) -> Dict[str, A
 def build_logger_config(
     path_parent: Optional[Path] = None,
     *, format_handler: Callable[[Path, bool], Dict[str, Any]] = _format_jsonl_handler,
-    production: bool = True
+    production: bool = True,
 ) -> Dict[str, Any]:
     """Build the loguru configuration. Use with `loguru.configure(**configuration)`.
 
@@ -171,7 +180,8 @@ def build_logger_config(
 
     Args:
         path_parent: Path to the directory where the '.logs/' folder should be created. Default is this package
-        production: if True, will tweak logging configuration for production code. Default is True
+        format_handler: function that will generate the loguru handler dictionary. Default is `_format_jsonl_handler`
+        production: if True, will tweak logging configuration for production code. Default is `True`
 
     Returns:
         Dict[str, Any]: the logger configuration as a dictionary