+++ noxfile.py
@@ -1,3 +1,3 @@
 """nox-poetry configuration file."""
 
-from calcipy.dev.noxfile import build_check, build_dist, check_safety, coverage, tests  # noqa: F401
+from calcipy.dev.noxfile import build_check, build_dist, check_safety, check_security, coverage, tests  # noqa: F401