+++ tests/test_doit_tasks/test_doc.py
@@ -9,6 +9,7 @@ from typing import List
 
 import pytest
 
+from calcipy.doit_tasks.base import write_text
 from calcipy.doit_tasks.doc import (
     _format_cov_table, _handle_coverage, _handle_source_file, _is_mkdocs_local,
     _move_cl, _parse_var_comment, task_cl_bump, task_cl_bump_pre, task_cl_write,
@@ -196,10 +197,12 @@ def test_task_document():
 
     assert _is_mkdocs_local() is False
     actions = result['actions']
-    assert len(actions) == 3
+    assert len(actions) == 5
     assert isinstance(actions[0][0], type(write_autoformatted_md_sections))
     assert str(actions[1]).startswith('Cmd: poetry run pdocs as_markdown')
-    assert str(actions[2]).startswith('Cmd: poetry run mkdocs build --site-dir')
+    assert isinstance(actions[2][0], type(write_text))
+    assert ' pyreverse ' in str(actions[3])
+    assert str(actions[-1]).startswith('Cmd: poetry run mkdocs build --site-dir')
 
 
 def test_task_open_docs():