+++ docs/docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 ## Unreleased
 
+### Feat
+
+- add code diagrams to doc-site
+- add notes on wily and pylint
+- add vulture
+- diff-quality (lint)
+- add it
+- add output if no stale packages found
+- add autoflake
+
+### Refactor
+
+- rename "wrappers" to "dot_dict"
+
+## 2021.0.3.2 (2021-06-06)
+
 ### Fix
 
 - error in noxfile with poetry install